intro-langchain-python/requirements.txt
# Dependências do projeto MBA AI Full-Cycle
# Gerado automaticamente - instalar com: pip install -r requirements.txt

# LangChain core packages
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9

# LangChain integrations
langchain-openai==0.3.31
langchain-google-genai==2.1.9

# Utilities
python-dotenv==1.1.1
beautifulsoup4==4.13.5
pypdf==6.0.0

# Core dependencies (automatically installed)
pydantic>=2.7.4
requests>=2.0.0
SQLAlchemy>=1.4.0